Simplify to server.js
